refactor(sample) : separate feature logics on ios
