+++ iosApp/iosApp/App/SampleFeature.swift
@@ -1,0 +1,15 @@
+enum SampleFeature: String, CaseIterable, Identifiable {
+    case textToSpeech
+    case speechToText
+
+    var id: Self { self }
+
+    var title: String {
+        switch self {
+        case .textToSpeech:
+            "Text to speech"
+        case .speechToText:
+            "Speech to text"
+        }
+    }
+}
