+++ iosApp/iosApp/Features/SpeechToText/TranscriptWord.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+struct TranscriptWord: Identifiable, Equatable {
+    let id: UUID
+    var text: String
+    var isVisible: Bool
+}
