+++ iosApp/iosApp/Features/TextToSpeech/TTSMode.swift
@@ -1,0 +1,8 @@
+enum TTSMode: String, CaseIterable, Identifiable {
+    case generate
+    case stream
+    case realtime
+
+    var id: Self { self }
+    var title: String { rawValue.capitalized }
+}
