+++ iosApp/iosApp/Shared/UI/ErrorView.swift
@@ -1,0 +1,24 @@
+import SwiftUI
+
+struct ErrorView: View {
+    let message: String
+    let actionTitle: String
+    let action: () -> Void
+
+    var body: some View {
+        VStack(alignment: .leading, spacing: 12) {
+            Text(message)
+                .foregroundStyle(.red)
+
+            Button(actionTitle, action: action)
+                .buttonStyle(.bordered)
+        }
+        .frame(maxWidth: .infinity, alignment: .leading)
+        .accessibilityElement(children: .contain)
+    }
+}
+
+#Preview {
+    ErrorView(message: "Something went wrong.", actionTitle: "Retry", action: {})
+        .padding()
+}
