--- iosApp/iosApp/iOSApp.swift
@@ -1,10 +1,0 @@
-import SwiftUI
-
-@main
-struct iOSApp: App {
-    var body: some Scene {
-        WindowGroup {
-            ContentView()
-        }
-    }
-}
